requirements.txt
python-dotenv==1.0.0
PyPDF2==3.0.1
langchain==0.1.14
langchain-community==0.0.34
langchain-text-splitters==0.0.1